requirements.txt
asgiref==3.6.0
black==23.3.0
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
decorator==5.1.1
Django==4.2
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.3.0
pycparser==2.21
pytz==2023.3
sqlparse==0.4.4
validators==0.20.0
